+++ src/test/java/pro/fessional/meepo/benchmark/PebbleBenchmark.java
@@ -1,6 +1,8 @@
 package pro.fessional.meepo.benchmark;
 
-import com.mitchellbosecke.pebble.error.PebbleException;
+import io.pebbletemplates.pebble.PebbleEngine;
+import io.pebbletemplates.pebble.error.PebbleException;
+import io.pebbletemplates.pebble.template.PebbleTemplate;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -15,7 +17,6 @@ import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
-import pro.fessional.meepo.sack.Gene;
 
 import java.io.IOException;
 import java.io.StringWriter;
@@ -33,21 +34,24 @@ import java.util.concurrent.TimeUnit;
 @BenchmarkMode(Mode.Throughput)
 @OutputTimeUnit(TimeUnit.SECONDS)
 @State(Scope.Benchmark)
-public class MeepoBenchmark {
+public class PebbleBenchmark {
 
-    private Gene template;
+    private PebbleTemplate template;
+
     private Map<String, Object> context;
 
     @Setup
     public void setup() throws PebbleException {
         context = Stock.mockContext();
-        template = pro.fessional.meepo.Meepo.parse("classpath:/template/jmh/stocks.meepo.html");
+        PebbleEngine engine = new PebbleEngine.Builder()
+                .autoEscaping(false).build();
+        template = engine.getTemplate("template/jmh/stocks.pebble.html");
     }
 
     @Benchmark
     public String benchmark() throws PebbleException, IOException {
         try (Writer writer = new StringWriter(10240)) {
-            template.merge(context, writer);
+            template.evaluate(writer, context);
             return writer.toString();
         }
     }
@@ -55,7 +59,7 @@ public class MeepoBenchmark {
     public static void main(String[] args) throws RunnerException {
         System.setProperty("org.slf4j.simpleLogger.log.pro.fessional.meepo", "error");
         Options opt = new OptionsBuilder()
-                .include(MeepoBenchmark.class.getSimpleName())
+                .include(PebbleBenchmark.class.getSimpleName())
                 .build();
 
         new Runner(opt).run();