+++ src/test/java/pro/fessional/meepo/tmpl/RnaTest.java
@@ -2,12 +2,18 @@ package pro.fessional.meepo.tmpl;
 
 import org.junit.jupiter.api.Test;
 import pro.fessional.meepo.TraceTest;
+import pro.fessional.meepo.poof.RnaManager;
+import pro.fessional.meepo.poof.impl.JsEngine;
 
 /**
  * @author trydofor
  * @since 2020-10-28
  */
 public class RnaTest extends TraceTest {
+
+    static {
+        RnaManager.register(new JsEngine());
+    }
     @Test
     public void putUse() {
         TmplHelp.assertTmpl("/template/rna/put-use-o.htm", "/template/rna/put-use-i.htm");